Added support for XDG_STATE_HOME
XDG_STATE_HOME as defined by the XDG Base Directroy Specification
is to store mutable files such as logs and other data that should
be preserved across applications.

As this is now a part of the XDG Base Directory Specification, it
is set to the expected default value of `$HOME/.local/state` unless
it already defined.

For info on this specification see:
https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html